requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM & Embeddings
anthropic>=0.18.0
openai>=1.12.0  # For embeddings

# Vector Database
pinecone-client>=3.0.0

# Document Processing
pypdf>=4.0.0
beautifulsoup4>=4.12.0
html2text>=2024.2.26
tiktoken>=0.6.0  # Token counting

# Data Processing
pandas>=2.2.0
python-multipart>=0.0.9  # For file uploads

# HTTP Client
httpx>=0.27.0
aiofiles>=23.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
